==> 02-Python-code/00-TestConnection.py <==
# comment 2
# comment 3
# comment 4
import json


filename = '/mnt/s3-refined-avvillas/Test-Amazon-Transcribe-Spanish-For-Voice-Analytics.json'
with open(filename, 'r') as f:
    datastore = json.load(f)
